optimize token expiry for socket listeners

+++ android/src/main/java/com/mosquitotransport/MosquitoTransportModule.kt
@@ -38,7 +38,7 @@ class MosquitoTransportModule(reactContext: ReactApplicationContext) :
   }
 
   override fun getSystemUptime(promise: Promise) {
-    val uptime = SystemClock.elapsedRealtime()
+    val uptime = SystemClock.elapsedRealtime().toDouble()
     promise.resolve(uptime)
   }
